Move Metric and its meta class in its own spec module and breaks a circular dependency
